+++ app/Http/Controllers/SpeciesController.php
@@ -51,10 +51,12 @@ class SpeciesController extends Controller
         $validatedData = $request->validated();
 
         $validatedData['interesting_cultivar'] = !isset($validatedData['interesting_cultivar']) ? null : explode(',', $validatedData['interesting_cultivar']);
+        $validatedData['suppliers'] = !isset($validatedData['suppliers']) ? null : $validatedData['suppliers'];
 
         $genus = Genus::find($validatedData['genus']);
         $species = $genus->species()->create($validatedData);
         $species->suppliers()->sync($validatedData['suppliers']);
+        $species->addMediaFromRequest('main_image')->toMediaCollection('main');
 
         return redirect()->route('species.index')->with('status', 'Added!');
     }